# app/Http/Controllers/API/RoleController.php
<?php

namespace App\Http\Controllers\API;

use Illuminate\Http\Request;
use Spatie\Permission\Models\Role;

class RoleController extends BaseController
{
    /**
     * @OA\Get (
     *     path="/auth/roles",
     *     tags={"Extra endpoints"},
     *     summary="Get list of roles",
     *     description="Returns list of roles",
     *     @OA\Response(
     *         response=200,
     *         description="Successful operation"
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     ),
     *     @OA\Response(
     *         response=403,
     *         description="Forbidden"
     *     )
     * )
     */
    public function index(Request $request)
    {
        return response()->json(Role::all());
    }
}